+++ src/debugger/variable/mod.rs
@@ -8,6 +8,7 @@ use std::fmt::{Display, Formatter};
 
 pub mod dqe;
 pub mod execute;
+pub mod mutability;
 pub mod render;
 pub mod value;
 pub mod r#virtual;